Change to always using MemoryStreamB, although StreamBaseByte can once again be used now. Always use replication for memory stream as memory stream now implements IDisposable. Standard usage is following:
use ms = new MemStream(...) // allocate a new MemStream
 use ms = Function(...) // if Function returns a MemStream

If function returns MemStream then use following to create MemStream to return (caller of function will then use: "use ms = Function(...)":

let ms = new MemStream(...)

If MemStream is passed to a function, then if function needs MemStream after it returns, it must replicate and keep a copy, for example:

use ms = new MemStream(...)
 ms.Write(...)
 queue.ToSend(cmd, ms)

Now, since queue.ToSend actually enqueues the memstream and needs to save it, it internally will create a replica.

MemoryStreamB replication is also now faster as both the buffers are RefCnt as well as the list of buffers is RefCnt, so replication does not need to now create a new List. It simply adds a refcnt to original list. As this is all internal, the user of MemoryStreamB does not need to take care of anything.