Fix several bugs and issues found during 30 node cluster sort benchmark test.
1. TCP implementation seems to have bug which causes RST to be sent back when sender and/or receiver
buffer sizes are manually adjusted.  Removing the manual setting, and reverting to default fixes this.
This seems like it may be due to either the receiver not reporting the correct flow control window and thus
sender overflows the receiver's buffer.

2. Simultaneous BeginConnect calls to large number of machines occaissionally causes some of them to fail, saying
that receiver did not respond in time.  This happens after about 15 seconds.  Issuing a second call to BeginConnect
succeeds in a few milliseconds.  This is somewhat of a hack solution, but solves the issue for now.

3. There was one bug which caused the job to fail the first time.  This was due to duplicate files in
AssemblyManagedDLL and JobDependencies.  When there was a duplicate file, the BlobFactory would check the hash
and the second call to triggerFunc would not execute (as BlobFactory only uses file hash as the key).  One
solution would be to also include TypeOf or other info such as name in the hash.  Alternatively, if the
hash is a duplicate of what is already in the BlobFactory, then it is reported as a "Duplicate" which then
sets status to "AllAvailable".

4. There was a race condition in certain cases causing dequeueItem to be null in Component.  This was caused by
GetOrAddProc calling InitMultipleProc prior to the Dequeue member being set.  Fixed by placing lock and only
calling InitMultipleProcess from within GetOrAddProc in cases when processing has already started.

Conflicts:
	samples/SortBenchmark/SortBenchmark.fs
	src/CoreLib/deploy.fs
	src/CoreLib/task.fs
	src/tools/tools/basenetwork.fs
	src/tools/tools/customserialize.fs